Move dashboard, update types, fix eslint linting, add estree to remove warning
