+++ src/app/dashboard/time-off/manage-requests/PendingTable.tsx
@@ -48,7 +48,9 @@ const PendingTable: React.FC = () => {
           </td>
           <td className="border-b border-gray-200 px-4 py-2">01/01/2024</td>
           <td className="border-b border-gray-200 px-4 py-2">01/05/2024</td>
-          <td className="border-b border-gray-200 px-4 py-2">It's a Girl!</td>
+          <td className="border-b border-gray-200 px-4 py-2">
+            It&apos;s a Girl!
+          </td>
           <td className="border-b border-gray-200 px-4 py-2">12/29/2023</td>
           <td className="border-b border-gray-200 px-4 py-2">yes or no</td>
         </tr>
