+++ src/app/page.tsx
@@ -5,13 +5,6 @@ export default async function Index() {
     <div className="flex w-full flex-1 flex-col items-center gap-20">
       <div className="animate-in flex max-w-4xl flex-1 flex-col gap-20 px-3 opacity-0">
         <SubNavBar selected="HR Services for Jane" />
-        <p>
-          Ideas for home page:
-          <ul className="list-disc">
-            <li>should immediately redirect to login if not logged in, otherwise serve as dashboard base</li>
-            <li>should be general info page</li>
-          </ul>
-        </p>
       </div>
     </div>
   );
